requirements.txt
gunicorn
flask

torch
torchvision
pillow
